requirements.txt
cryptography==42.0.4
bcrypt==4.1.2
requests==2.31.0
ipinfo==5.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
python-dotenv==1.0.1
pydantic==1.10.13
resend==0.7.2
pymongo==4.6.3
typeddict==0.3.0
pyotp==2.9.0
sentry-sdk==2.16.0
fastapi==0.115.6